ZEnet/requirements.txt
annotated-doc==0.0.4
click==8.3.1
irc==20.5.0
jaraco.collections==5.2.1
jaraco.context==6.1.0
jaraco.functools==4.4.0
jaraco.logging==3.4.0
jaraco.stream==3.0.4
jaraco.text==4.2.0
markdown-it-py==4.0.0
mdurl==0.1.2
more-itertools==10.8.0
mysql-connector-python==9.6.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
rich==14.3.2
shellingham==1.5.4
six==1.17.0
tempora==5.8.1
typer==0.23.1
typer-slim==0.23.1
